Move dates into its own directory
